-removed a redundant link in built

+++ saul-examples/src/main/scala/edu/illinois/cs/cogcomp/saulexamples/BasicClassifierWithRandomData/randomDataModel.scala
@@ -21,7 +21,7 @@ object randomDataModel extends DataModel {
 
   val randomProperty = property(randomNode, cache = true) {
     x: String =>
-      val p = List(2 * r.nextGaussian().toDouble, 0.5 * r.nextGaussian().toDouble)
+      val p = List(2 * r.nextGaussian(), 0.5 * r.nextGaussian())
       val p1new = p(1) + randomLabel(x).toDouble
       val p2 = List(c * p1new - s * p(1), c * p1new + s * p(1))
       p2
